+++ main.py
@@ -1,4 +1,4 @@
-__version__ = "0.2.1a"
+__version__ = "0.2.2a"
 
 from bootstrap import main